Massive reduction in bundled parser size.
Reduced the size of the parser from 416.89 kB to 86.7 kB
(about 20% of the original size) by switching pegjs option
--optimize from speed to size and modifying my fork of
pegjs to allow rule descriptions to be looked up by rule
index instead of by rule name as the optimize size mode
required.

Running time for the entire suite of specs is still around
1 second, so the tracer branch is now performing as fast
as master and is also a fraction of the file size as well.

Refs #16