Do not treat a transaction as a single statement.
Now, instead of transaction statements being parsed as a single
statement of type `transaction` to be considered valid, each
statement that makes up a the transaction (e.g., `BEGIN`,
`END`) is considered its own distinct statement that can exist
independent of the others. Because a single transaction can be
spread across multiple input strings given to the parser, it is no
longer treated as a single, large, _transaction_ statement.

  ``` sql
  BEGIN;
  DROP TABLE t1;
  END;
  ```

  ``` json
  {
    "type": "statement",
    "variant": "list",
    "statement": [
      {
        "type": "statement",
        "variant": "transaction",
        "action": "begin"
      },
      {
        "type": "statement",
        "target": {
          "type": "identifier",
          "variant": "table",
          "name": "t1"
        },
        "variant": "drop",
        "format": "table",
        "condition": []
      },
      {
        "type": "statement",
        "variant": "transaction",
        "action": "commit"
      }
    ]
  }
  ```