Add --stream option to CLI. Refs #26
Automatically switch to streaming mode when the input file is
over 150kB or manually enable with --stream or -x.

Not made the default mode since it is actually slower for files
that can be parsed using the default method without running out
of available memory. The hard limit for that is around 200kB on
my machine.

Refs #27